+++ CHANGELOG.md
@@ -7,12 +7,19 @@ The format of this changelog is based on ["Keep a Changelog"](http://keepachange
 This project adheres to [Semantic Versioning](http://semver.org/). Version numbers follow the pattern: `MAJOR.FEATURE.BUILD`
 
 
+## 3.0.0-beta.3 - 2020-01-08
+
+### Changed
+
+- Refactored the logic for instantiating element queries. (You can now directly specify your element and query classes of choice.)      
+
+
 ## 3.0.0-beta.2 - 2018-10-22
 
 ### Changed
 
 - Updated the `maxPowerCaptain` call to use the current Craft 3 `App` helper.
-- Rewrote the task infrastructure to use the new Craft 3 queue/jobs system.  
+- Rewrote the task infrastructure to use the new Craft 3 queue/jobs system.
 - Re-released under the Craft license.